Add liveblog structured data for dotcom-rendering
Note, as part of this, the linked data code for DCR has been
separated out. This has resulted in a little duplication of a few
modules (e.g. Publisher) but the benefit is to isolate the DCR-specific
code, which should make future cleanup easier.
